net: check for LXC_NET_EMPTY/NONE instead of zero ifindex
This is to explicitly state the intention of this section
via the code since it previously covered to independent
cases. These were type "empty" interfaces first, and later
on veths instantiated when starting as unprivileged user.
Since the latter caused other invalid configurations to be
silently accepted and misinterpreted a stricter approach at
error handling should be useful.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>