net: error trying to assign invalid network interfaces
The only interfaces allowed to have a 0 index are
LXC_NET_EMPTY, LXC_NET_NONE as well as veths instantiated
for unprivileged users via lxc-user-nic. The latter case is
already handled above, but the interface index can also be
zero in more cases such as when an unprivileged user tries
to assign physical interfaces. The instantiation callbacks
aren't executed and thus the ifindex is not filled in.
Instead of silently accepting this error out.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>