net: lxc-user-nic should print the container-side ifindex
In privileged startups the instantiation functions fill
netdev->ifindex while in unprivileged startups only veth
devices are currently supported and the index is left as
zero to be figured out in setup_netdev(). This clashes with
the LXC_NET_EMPTY case and causes other device types to be
silently accepted and interpreted as "empty"-typed devices
during setup.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>